requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.2
djangorestframework==3.16.0
joblib==1.4.2
nltk==3.9.1
regex==2024.11.6
sqlparse==0.5.3
tqdm==4.67.1
